Fix finding of bibliography.
Previously, LaTeX could find bibliographies anywhere in the LaTeX path,
but pandoc's CSL could not. So to use CSL, complete paths would be
required. However, this breaks when documents are put on different
systems that place LaTeX bibliographies in different folders (as between
MacOS and Linux). These changes fix that by using `kpsewhich` to locate
LaTeX bibliographies, and inserting the found location into the document
for use by CSL if needed.